refactor: modulurize blobs storage abstractions
